Correction of router code and added path to controller
