--- routers/me.router.js
@@ -1,15 +1,0 @@
-const express = require("express"); 
-const morgan = require("morgan"); 
-
-const booksRouter = express.Router(); 
-
-booksRouter.use(morgan("dev")); 
-
-booksRouter.get("/books"); 
-booksRouter.get("/books/:id"); 
-booksRouter.post("/books"); 
-booksRouter.put("/books/:id"); 
-booksRouter.patch("/books/:id"); 
-booksRouter.delete("/books/:id"); 
-
-module.exports = booksRouter;
